+++ packages/ranger-apollo/src/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useAwaitQuery } from './useAwaitQuery'