+++ packages/ranger-apollo/src/index.ts
@@ -1,0 +1,2 @@
+export { shrinkFetchQuery, shrinkQuery } from './graphql'
+export { useAwaitQuery } from './hooks'