+++ packages/ranger-contentful-render/src/apollo/index.ts
@@ -1,0 +1,2 @@
+export type { ContentfulType } from './createLink'
+export { createContentfulLink } from './createLink'