+++ packages/ranger-contentful-render/src/components/index.ts
@@ -1,0 +1,2 @@
+export type { RichTextRenderProps } from './richTextRender'
+export { default as RichTextRender } from './richTextRender'