+++ packages/ranger-core/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @ranger-theme/core
 
+## 0.12.0
+
+### Minor Changes
+
+- publish the contentful and apollo
+
+### Patch Changes
+
+- Updated dependencies
+  - @ranger-theme/apollo@0.1.0
+
 ## 0.11.0
 
 ### Minor Changes